Tests/LibWeb/Text/input/css/counter-style-system-fixed-setter.html

<!doctype html>
<style>
    @counter-style foo {
        system: fixed;
        symbols: "A" "B";
    }
</style>
<script src="../include.js"></script>
<script>
    test(() => {
        const rule = document.styleSheets[0].cssRules[0];
        println(rule.system);

        rule.system = "fixed 1";
        println(rule.system);

        rule.system = "fixed 2";
        println(rule.system);
    });
</script>
