Tests/LibWeb/Text/expected/css/counter-style-system-fixed-setter.txt

fixed
fixed 1
fixed 1
